subindo

+++ src/pages/Funcionarios/Funcionarios.jsx
@@ -4,7 +4,7 @@ import Pesquisa from '../../components/Pesquisa/Pesquisa';
 import CardGeral from '../../components/Cards/CardGeral';
 import { useNavigate } from 'react-router';
 import { useEffect, useState } from 'react';
-import Button from 'react-bootstrap/Button'; 
+import Button from 'react-bootstrap/Button';
 import { FaEdit, FaRegTrashAlt } from 'react-icons/fa';
 
 const Funcionarios = () => {
@@ -38,16 +38,16 @@ const Funcionarios = () => {
     fetch(`http://localhost:3000/deletarFuncionario/${id}`, {
       method: 'DELETE'
     })
-    .then(response => {
-      if(!response.ok){
-        throw new Error('Erro ao deletar o funcionário')
-      }
-      console.log('Funcionário deletado com sucesso')
-      window.location.reload();
-    })
-    .catch(error => {
-      console.error(error)
-    })
+      .then(response => {
+        if (!response.ok) {
+          throw new Error('Erro ao deletar o funcionário')
+        }
+        console.log('Funcionário deletado com sucesso')
+        window.location.reload();
+      })
+      .catch(error => {
+        console.error(error)
+      })
   }
 
 
@@ -81,31 +81,36 @@ const Funcionarios = () => {
           imgHeight={250}
           onCardClick={handleCardClick}
           showButtons={false}
-          customButton={ item => (
+          customButton={item => (
             <>
               <Button
-                      variant='warning'
-                      className='rounded-circle fs-5 text-center shadow m-1'
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        if (item.acoes && item.acoes[0]?.onClick) item.acoes[0].onClick();
-                      }}
-                    >
-                      <FaEdit />
-                    </Button>
-                    <Button
-                      variant='danger'
-                      className='rounded-circle fs-5 text-center shadow m-1'
-                      onClick={(e) => {
-                        e.stopPropagation();
-                        handleDelete(item.id);
-                        if (item.acoes && item.acoes[1]?.onClick) item.acoes[1].onClick();
-                      }}
-                    >
-                      <FaRegTrashAlt />
-                    </Button>
+                variant='warning'
+                className='rounded-circle fs-5 text-center shadow m-1'
+                onClick={(e) => {
+                  e.stopPropagation();
+                  if (item.acoes && item.acoes[0]?.onClick) item.acoes[0].onClick();
+                }}
+              >
+                <FaEdit />
+              </Button>
+              <Button
+                variant='danger'
+                className='rounded-circle fs-5 text-center shadow m-1'
+                onClick={(e) => {
+                  e.stopPropagation();
+                  {
+                    const confirmar = window.confirm('Deseja deletar o funcionário?')
+                    if (confirmar) {
+                      handleDelete(item.id);
+                    }
+                  } 
+                  if (item.acoes && item.acoes[1]?.onClick) item.acoes[1].onClick();
+                }}
+              >
+                <FaRegTrashAlt />
+              </Button>
             </>
-             )
+          )
           }
         />
       </Container>
